refactor: 안쓰는 클래스 삭제
