+++ src/main/java/com/beauty4u/backend/inform/command/application/service/InformService.java
@@ -1,7 +1,5 @@
 package com.beauty4u.backend.inform.command.application.service;
 
-import com.beauty4u.backend.file.command.domain.service.FileDomainService;
-import com.beauty4u.backend.common.util.S3ImageUtil;
 import com.beauty4u.backend.inform.command.application.dto.InformReqDTO;
 import com.beauty4u.backend.inform.command.application.dto.UpdateInformViewcount;
 import com.beauty4u.backend.inform.command.domain.service.InformDomainService;
@@ -14,8 +12,6 @@ import org.springframework.transaction.annotation.Transactional;
 public class InformService {
 
     private final InformDomainService informDomainService;
-    private final FileDomainService fileDomainService;
-    private final S3ImageUtil s3ImageUtil;
 
     @Transactional
     public Long saveInform(String loginUserCode, InformReqDTO informReqDTO) {
