+++ src/main/java/com/beauty4u/backend/inform/command/domain/service/InformDomainService.java
@@ -2,7 +2,6 @@ package com.beauty4u.backend.inform.command.domain.service;
 
 import com.beauty4u.backend.common.exception.CustomException;
 import com.beauty4u.backend.common.exception.ErrorCode;
-import com.beauty4u.backend.common.util.S3ImageUtil;
 import com.beauty4u.backend.inform.command.application.dto.InformDTO;
 import com.beauty4u.backend.inform.command.application.dto.InformReqDTO;
 import com.beauty4u.backend.inform.command.application.dto.UpdateInformViewcount;
@@ -10,11 +9,9 @@ import com.beauty4u.backend.inform.command.domain.aggregate.Inform;
 import com.beauty4u.backend.inform.command.domain.repository.InformRepository;
 import com.beauty4u.backend.user.command.domain.aggregate.UserInfo;
 import com.beauty4u.backend.user.command.domain.repository.UserRepository;
-import com.beauty4u.backend.user.command.domain.service.UserDomainService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
@@ -22,9 +19,7 @@ public class InformDomainService {
 
     private final InformRepository informRepository;
     private final UserRepository userRepository;
-    private final S3ImageUtil s3ImageUtil;
     private final ModelMapper modelMapper;
-    private final UserDomainService userDomainService;
 
     public Long saveInform(String loginUserCode, InformReqDTO informReqDTO) {
 
@@ -56,9 +51,6 @@ public class InformDomainService {
     }
 
     public void deleteInform(Long informId) {
-
-        Inform inform = informRepository.findById(informId)
-                .orElseThrow(() -> new CustomException(ErrorCode.NOT_FOUND_INFORM));
 
         informRepository.deleteById(informId);
     }
