--- src/main/java/com/beauty4u/backend/teamspace/command/application/service/ScrapService.java
@@ -1,97 +1,0 @@
-package com.beauty4u.backend.teamspace.command.application.service;
-
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.model.CannedAccessControlList;
-import com.amazonaws.services.s3.model.ObjectMetadata;
-import com.amazonaws.services.s3.model.PutObjectRequest;
-import com.beauty4u.backend.teamspace.command.domain.aggregate.Folder;
-import com.beauty4u.backend.teamspace.command.domain.repository.FolderRepository;
-import com.beauty4u.backend.teamspace.command.domain.repository.ScrapRepository;
-import com.beauty4u.backend.user.command.domain.aggregate.UserInfo;
-import com.beauty4u.backend.user.command.domain.repository.UserRepository;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.util.UUID;
-
-@Service
-@RequiredArgsConstructor
-@Transactional
-@Slf4j
-public class ScrapService {
-
-    private final AmazonS3 amazonS3;
-    private final FolderRepository folderRepository;
-    private final UserRepository userRepository;
-
-    @Value("${cloud.aws.s3.bucket}")
-    private String bucketName;
-
-    public String uploadScrapImage(Long teamspaceId, Long folderId, MultipartFile file) {
-
-        // 폴더 조회
-        Folder folder = folderRepository.findById(folderId)
-                .orElseThrow(() -> new IllegalArgumentException("Folder not found"));
-
-        // 사용자 조회
-        UserInfo user = userRepository.findByUserCode("user-code-placeholder")
-                .orElseThrow(() -> new IllegalArgumentException("User not found"));
-
-        // 파일 업로드 처리
-        String fileUrl = uploadFileToS3(file);
-        return fileUrl;
-
-    }
-
-    private String uploadFileToS3(MultipartFile file) {
-        String fileName = generateUniqueFileName(file.getOriginalFilename());
-
-        // 메타데이터 설정
-        ObjectMetadata metadata = new ObjectMetadata();
-        metadata.setContentType(file.getContentType());
-        metadata.setContentLength(file.getSize());
-
-
-        try {
-            amazonS3.putObject(new PutObjectRequest(bucketName, fileName, file.getInputStream(), metadata)
-                    .withCannedAcl(CannedAccessControlList.PublicRead));
-        } catch (IOException e) {
-            throw new RuntimeException("Failed to upload file to S3", e);
-        }
-
-        return amazonS3.getUrl(bucketName, fileName).toString();
-    }
-
-    private String generateUniqueFileName(String originalFilename) {
-        String uuid = UUID.randomUUID().toString();
-        return uuid + "_" + originalFilename;
-    }
-
-
-
-    // 테스트용
-    public String uploadImage(MultipartFile image) throws IOException {
-        String fileName = UUID.randomUUID() + "_" + image.getOriginalFilename(); // 고유한 파일 이름 생성
-
-        // 메타데이터 설정
-        ObjectMetadata metadata = new ObjectMetadata();
-        metadata.setContentType(image.getContentType());
-        metadata.setContentLength(image.getSize());
-
-        // S3에 파일 업로드 요청 생성
-        PutObjectRequest putObjectRequest = new PutObjectRequest(bucketName, fileName, image.getInputStream(), metadata);
-
-        // S3에 파일 업로드
-        amazonS3.putObject(putObjectRequest);
-
-        return String.format("https://%s.s3.%s.amazonaws.com/%s", bucketName, amazonS3.getRegionName(), fileName);
-
-    }
-}
-
-
