--- src/main/java/com/beauty4u/backend/teamspace/command/domain/aggregate/FolderStatus.java
@@ -1,8 +1,0 @@
-package com.beauty4u.backend.teamspace.command.domain.aggregate;
-
-import lombok.Getter;
-
-@Getter
-public enum FolderStatus {
-    ACTIVE, INACTIVE
-}
