--- src/main/java/com/beauty4u/backend/teamspace/command/domain/aggregate/ScrapStatus.java
@@ -1,5 +1,0 @@
-package com.beauty4u.backend.teamspace.command.domain.aggregate;
-
-public enum ScrapStatus {
-    ACTIVE, INACTIVE
-}
