--- src/main/java/com/beauty4u/backend/teamspace/command/domain/repository/ScrapRepository.java
@@ -1,10 +1,0 @@
-package com.beauty4u.backend.teamspace.command.domain.repository;
-
-import com.beauty4u.backend.teamspace.command.domain.aggregate.Scrap;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ScrapRepository extends JpaRepository<Scrap, Long> {
-
-}
