+++ src/main/java/com/beauty4u/backend/teamspace/query/controller/ChatRoomQueryController.java
@@ -1,5 +1,8 @@
 package com.beauty4u.backend.teamspace.query.controller;
 
+import com.beauty4u.backend.common.response.ApiResponse;
+import com.beauty4u.backend.common.response.ResponseUtil;
+import com.beauty4u.backend.common.success.SuccessCode;
 import com.beauty4u.backend.common.util.CustomUserUtil;
 import com.beauty4u.backend.teamspace.query.dto.chatroom.ChatRoomDTO;
 import com.beauty4u.backend.teamspace.query.dto.chatroom.ChatRoomDetailsDto;
@@ -23,29 +26,36 @@ public class ChatRoomQueryController {
 
     @Operation(summary = "채팅방 리스트 조회", description = "본인이 속한 채팅방 리스트를 조회합니다.")
     @GetMapping("/rooms")
-    public ResponseEntity<List<ChatRoomDTO>> getMyChatRooms() {
+    public ResponseEntity<ApiResponse<List<ChatRoomDTO>>> getMyChatRooms() {
 
         String loginUserCode = CustomUserUtil.getCurrentUserCode();
         List<ChatRoomDTO> chatRooms = chatRoomQueryService.getMyChatRooms(loginUserCode);
-        return ResponseEntity.ok(chatRooms);
+
+        return ResponseUtil.successResponse(SuccessCode.CHAT_ROOM_LIST_FIND_SUCCESS, chatRooms);
     }
 
     @Operation(summary = "채팅방 내의 채팅 참여자 정보 조회", description = "부서 코드에 맞는 채팅방 참여자의 정보를 조회합니다.")
     @GetMapping("/{chatRoomId}/users")
-    public List<ChatRoomUserInfoDto> getChatMemberByChatRoomId(
+    public ResponseEntity<ApiResponse<List<ChatRoomUserInfoDto>>> getChatMemberByChatRoomId(
             @PathVariable Long chatRoomId) {
 
         String loginUserCode = CustomUserUtil.getCurrentUserCode();
-        return chatRoomQueryService.findAllChatRoomUser(chatRoomId, loginUserCode);
+
+        List<ChatRoomUserInfoDto> allChatRoomUser = chatRoomQueryService.findAllChatRoomUser(chatRoomId, loginUserCode);
+
+        return ResponseUtil.successResponse(SuccessCode.CHAT_ROOM_USER_LIST_FIND_SUCCESS, allChatRoomUser);
     }
 
 
     // 팀스페이스 통합 컨트롤러
     @Operation(summary = "채팅방 채팅 상세조회", description = "채팅방 채팅방 참여자, 채팅 목록을 조회합니다.")
     @GetMapping("/{chatRoomId}/details")
-    public ChatRoomDetailsDto getChatDetails(@PathVariable Long chatRoomId) {
+    public ResponseEntity<ApiResponse<ChatRoomDetailsDto>> getChatDetails(@PathVariable Long chatRoomId) {
 
         String loginUserCode = CustomUserUtil.getCurrentUserCode();
-        return chatRoomQueryService.getChatRoomDetails(chatRoomId, loginUserCode);
+
+        ChatRoomDetailsDto chatRoomDetails = chatRoomQueryService.getChatRoomDetails(chatRoomId, loginUserCode);
+
+        return ResponseUtil.successResponse(SuccessCode.CHAT_ROOM_DETAIL_FIND_SUCCESS, chatRoomDetails);
     }
 }
