+++ src/main/java/com/beauty4u/backend/user/query/controller/UserQueryController.java
@@ -4,7 +4,11 @@ import com.beauty4u.backend.common.response.ApiResponse;
 import com.beauty4u.backend.common.response.ResponseUtil;
 import com.beauty4u.backend.common.success.SuccessCode;
 import com.beauty4u.backend.common.util.CustomUserUtil;
-import com.beauty4u.backend.user.query.dto.*;
+import com.beauty4u.backend.user.query.dto.dept.DeptResDTO;
+import com.beauty4u.backend.user.query.dto.job.JobResDTO;
+import com.beauty4u.backend.user.query.dto.role.RoleResDTO;
+import com.beauty4u.backend.user.query.dto.user.FindUserDetailResDTO;
+import com.beauty4u.backend.user.query.dto.user.UserListResDTO;
 import com.beauty4u.backend.user.query.service.UserQueryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
