+++ src/main/java/com/beauty4u/backend/user/query/dto/job/JobResDTO.java
@@ -1,4 +1,4 @@
-package com.beauty4u.backend.user.query.dto;
+package com.beauty4u.backend.user.query.dto.job;
 
 import lombok.Getter;
 import lombok.Setter;
