+++ src/main/java/com/beauty4u/backend/user/query/dto/role/RoleResDTO.java
@@ -1,4 +1,4 @@
-package com.beauty4u.backend.user.query.dto;
+package com.beauty4u.backend.user.query.dto.role;
 
 import lombok.Getter;
 import lombok.Setter;
