+++ src/main/java/com/beauty4u/backend/user/query/mapper/DeptQueryMapper.java
@@ -1,6 +1,6 @@
 package com.beauty4u.backend.user.query.mapper;
 
-import com.beauty4u.backend.user.query.dto.DeptResDTO;
+import com.beauty4u.backend.user.query.dto.dept.DeptResDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
