+++ src/main/java/com/beauty4u/backend/user/query/mapper/JobQueryMapper.java
@@ -1,6 +1,6 @@
 package com.beauty4u.backend.user.query.mapper;
 
-import com.beauty4u.backend.user.query.dto.JobResDTO;
+import com.beauty4u.backend.user.query.dto.job.JobResDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
