+++ src/main/java/com/beauty4u/backend/user/query/mapper/RoleQueryMapper.java
@@ -1,6 +1,6 @@
 package com.beauty4u.backend.user.query.mapper;
 
-import com.beauty4u.backend.user.query.dto.RoleResDTO;
+import com.beauty4u.backend.user.query.dto.role.RoleResDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
