--- src/main/resources/mapper/teamSpace/folder/FolderMapper.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.beauty4u.backend.chatRoom.query.mapper.FolderQueryMapper">
-
-    <select id="findFolderList" resultType="FolderListResDTO">
-        SELECT f1.folder_id,
-               f1.top_folder_id,
-               f1.teamspace_id,
-               f1.folder_name
-        FROM folder f1
-                 LEFT JOIN folder f2 ON f1.top_folder_id = f2.folder_id
-        WHERE f1.teamspace_id = #{chatRoomId}
-          AND f1.publish_status = #{folderStatus}
-        <choose>
-            <when test="topFolderId == null">
-                and f1.top_folder_id IS NULL
-            </when>
-            <otherwise>
-                and f1.top_folder_id = #{topFolderId}
-            </otherwise>
-        </choose>
-
-        <if test="folderName != null and folderName != ''">
-            and f1.folder_name LIKE CONCAT('%', #{folderName}, '%')
-        </if>
-        ORDER BY f1.folder_name
-    </select>
-</mapper>
